update (i3): new compositor and file manager
Picom has lot of great features and doing it's work very well but it's little bit bloated.
I was searching for a effiecent compositor for daily ussage and I found compton which is
successor of picom. So, anymore I'm gonna use comptoon as a desktop compositor.
I was using Thunar more than 8-years and it is stable as rock. It's
feature complete but it has limited features so I need some plugins for ftp, samba, archives etc.
So, I want to giva a chance to Dolphin which has lots of more feature built-in.

Picom -> Compton
Thunar -> Dophin